# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "API documentation for my Laravel project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/company-services": {
            "get": {
                "tags": [
                    "Company Service"
                ],
                "summary": "Fetch all company services",
                "operationId": "0762eab673099bb6cdfedff5e823e6f1",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Company Service"
                ],
                "summary": "Store a new company service",
                "operationId": "78ab3e544b1344a22cc1a0d85e2f773d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "companyServiceName",
                                    "companyServiceImageUrl"
                                ],
                                "properties": {
                                    "companyServiceName": {
                                        "description": "Name of the company service",
                                        "type": "string"
                                    },
                                    "companyServiceImageUrl": {
                                        "description": "Image URL of the company service",
                                        "type": "string"
                                    },
                                    "serviceItemIds": {
                                        "description": "Array of service item IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company service created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company-services/{id}": {
            "get": {
                "tags": [
                    "Company Service"
                ],
                "summary": "Show a single company service",
                "operationId": "1ca265e74461c7bd461443ca805323d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Company service not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Company Service"
                ],
                "summary": "Update a company service",
                "operationId": "aba6d10eaf6d90e3752546ae15d1c7fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "companyServiceName": {
                                        "description": "Name of the company service",
                                        "type": "string"
                                    },
                                    "companyServiceImageUrl": {
                                        "description": "Image URL of the company service",
                                        "type": "string"
                                    },
                                    "serviceItemIds": {
                                        "description": "Array of service item IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company service updated successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "404": {
                        "description": "Company service not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Company Service"
                ],
                "summary": "Delete a company service",
                "operationId": "38cefece89b9fa3506b42e35967f4358",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company service deleted successfully"
                    },
                    "404": {
                        "description": "Company service not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}